Changed price in Product model to PositiveIntegerField to prevent negative or blank values being entered editing a product

+++ products/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
+from django.core.validators import MaxValueValidator
 
 # Create your models here.
 
@@ -27,7 +28,7 @@ class Product(models.Model):
     description = models.TextField()
     has_sizes = models.BooleanField(default=False, null=True, blank=True)
     string_gauge = models.BooleanField(default=False, null=True, blank=True)
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.PositiveIntegerField()
     rating = models.DecimalField(max_digits=6, decimal_places=2, null=True,
                                  blank=True)
     image_url = models.URLField(max_length=1024, null=True, blank=True)
